Add an option to skip site links
Site links take up a ton of space so it might be worth skipping them.
IDs 1-200 (including the missing IDs) is ~60,000 triple without site links
and ~150,000 with them.

Also fix an error encountered loading string with \ in them.  This would
be an isolated commit but its test overlaps with the above and it isn't
worth it to break it out.

Change-Id: Icc8a3b4c327e4719c0eb221060218448ebcdd877