Support for complex statements
This means that complex statements (statements who's objects aren't values
but instead are IRIs that have many values) are properly munged and properly
synced.  The most interesting part of this is handling shared references.
Wikibase gives shared references the same IRIs so we need to make sure that
they don't get cleared when they are removed from a single node.

Change-Id: I8d35526bbd6883ff6b13f0f1dfa610f8fd73d2a6