Remove timeout tests
Tests that verify that things run quickly with the timeout parameter are
prone to give false negatives on slow VMs.  So this patch rewrites them with
weaker assertions - instead of asserting speed they assert that things like
excess mutations are not performed.  Its less strong but its liable to make
the build more stable.

Change-Id: Iddfe50313694232b63dadeaae166a62ca39c2276