Make sure property edits show up in recent changes
For a while we were polling for recent changes just in the main namespace
but properties are in the property namespace. This polls both namespaces now.

Also we now take care not to create a new page with every test run.  Now
we only create new pages on the first run and reuse it afterwards.

Change-Id: Iccd2f8a597439f3c9f3002cf0bcebef101c781df