Support munging dump into chunks
The Munge tool now has a --chunkSize option which supports spitting out
files chunked on some number of entities.

Change the --to port:nnnn behavior so that the entire dump is only generated
once.  Without --chunkSize that means that the process exits after serving a
single http request.  With --chunkSize that means every response is the next
chunk.

Also fix an error in the mormalizer that was leaving the versioned ontology
in.

Change-Id: If8e348a0089640380e774c12f2085d0dfefe50d4