singleLable option shouldn't make up labels
Originally the idea behind the single label option was that the munge process
would always output only a single label for each entity so you could always
just add a ```?e rdfs:label ?eLabel``` clause to get the label. This sounds
good but to do it you have to add labels to entities that don't have labels
in any of the languages in the fallback chain and thats silly because those
labels are fake. So instead we just don't add any labels if there aren't any
in the fallback chain.  The only disadvantage here is that when searching the
graph you have to add a ```OPTIONAL { ?e rdfs:label ?eLabel }``` clause to get
the label. That isn't that big a deal to type anyway and we can reduce the
typing using magic syntax if we need to.

Closes T95779

Change-Id: Ic710816fd577c67a26b97f9a7c7415b65acf6ede