Add a mechanism to bundle for distribution
Fixes #T96148

This adds a new subproject that bundles everything into a nice little
package.  It also avoids the need to *mvn install* anything, as it gets
all its project references from the build, rather than from the local
Maven repo.

This makes redundant (and likely deprecates) the similar infrastructure
in *tools* which extracts the Blazegraph war, copies in the project
artifacts, and the shell scripts.  I've left that stuff alone for now,
since I'm not 100% certain it's no longer needed.  Either way, we can
prune it later as appropriate.

Change-Id: I144ecf4433c05e9223f208d7dec31d511eca13e5