Setup checkstyle
Also add a ton of Javadoc to get them to pass.

And some cleanup to get them to pass as well.

Some of the rules we use in checkstyle are a bit debatable - but its nice to
have some set of rules. We can always change them if we don't like them.

I left a few instances of high class "fan out complexity" in the code because
I don't have time to rework them and the commit is already plenty big enough.

You can skip running checkstyle and the forbidden apis check by running mvn
with `-PnoVerify`.

Change-Id: I4a6d2142507121283d3c05a70d81df8a00c72bfc