Resolve labels
Implements a "service" that resolves label like things in a way that doesn't
change the cardinality of the result set. You can call it like this:
```
 SELECT *
 WHERE {
   SERVICE wikibase:label.en.de.fr {
     wd:Q123 rdfs:label ?q123Label .
     wd:Q123 rdfs:altLabel ?q123Alt .
     wd:Q123 schema:description ?q123Desc .
     wd:Q321 rdf:label ?q321Label .
   }
 }
```
or like this:
```
 SELECT ?sLabel ?sAltLabel ?sDescription ?oLabel
 WHERE {
   ?s wdt:P22 ?o .
   SERVICE wikibase:label.en.de.fr {
   }
 }
```

It works by resolving the labels one by one on each solution. Blazegraph
naturally pushes SERVICE calls to the last step in the query which is
perfect for this.

Change-Id: I37df29ad442692d350fdd35cd17badb05170b416