Add secondary poller which lags behind and tries to collect missed updates.
The poller (enabled by -T <seconds> option to Updater) runs
as a separate thread and tries to identify missed updates and collect them.

The code also stored all seen RC IDs for several hours back to know which
updates where seen and which weren't. Currently it's 360000 IDs back,
which should be enough for now.

Change-Id: I21a77f4d32b903dac4d9b1a842d8af78d6f9c5ed