Adding findbugs to the build
Findbugs violations will fail the build. All current violations have been
corrected or marked with @SuppressFBWarnings. Most are minor, but a few
could be improved. In particular, moving a the new java.time.LocalDateTime
instead of java.util.Date would be cleaner. This might come in another change
since it is non trivial to implement.

Change-Id: Id5886aece3bfbe83b308a58f1d9d115ddeedfb5b