deploy blazegraph only once
Copying the blazegraph-service war to the tools module and start
it explicitely. Using a test dependency would mess up the test
classpath (see note below). The runBlazegraph.sh script has moved
to the war module (where it logically belongs) to ensure that it
can run standalone, without requiring the copy-war execution to
be run first.

Note: tests were failing due to the misaligned versions of
sesame. This should be fixed at some point.
Change-Id: I1b9984c45d2f2d076c7205fe0e3f43bebf31886f