Send logs to logstash
Add logstash-logback-encoder as a dependency to enable sending logs to
logstash. The acutal logback configuration is entirely in puppet. This
repo provides a default logback configuration which is loaded unless
a different configuration is specified in /etc/default/wdqs-*

This change needs to be deployed with its related puppet change.

Bug: T172710
Change-Id: Ie9075dfbc34055748976648fa3d06a91fd2d9563