Make WikibaseRepository type safe
Parse responses from Wikibase to Java objects to make them type safe. This
makes the client code less error prone, contains the JSON parser dependency
to the WikibaseRepository class and makes refactoring those classes easier.

This replaces json-simple with Jackson, which is already a dependency in
the project.

Still to check:

At this point, an invalid change will put the whole batch of changes in
error. This could happen if a field is invalid (non parseable date in a
date field, invalid integer in an integer field, ...). I'm not sure if
that's an error that is sufficiently frequent to need specific handling.

Change-Id: Ibd2a2e217931333dafa894dd003f3fdbf6c2157e