update jackson to version 2.9.3
The version of jackson that we are using is outdated (> 4 years) and will
cause issues when introducing wiremock (or other lib requiring a reasonably
recent version of jackson). And it is most of the time a good idea to keep
dependencies up to date.

Tests are still passing with this new version and jackson has a good track
record of preserving compatibility.

Change-Id: I4051bd08872a54284796a49383346fb14a00f147