Set npm version for frontend-maven-plugin
When using NodeJS 4+, the frontend-maven-plugin uses the bundled npm,
but it ends up not being available via PATH lookup. That causes issue
when one package.json has a script that invokes npm (it can not be
found).

Set <npmVersion> to the version being used on Wikimedia CI which would
hopefully install npm in a directory belonging to PATH.

https://github.com/eirslett/frontend-maven-plugin/issues/560
https://github.com/eirslett/frontend-maven-plugin#installing-node-and-npm

Bug: T188285
Change-Id: Ia892dcfdc1c0ef055f00a493941e974198323012