Extracting RdfClient from RdfRepository.
This RdfClient handles the low level logic of connecting to blazegraph,
while the higher level logic and actual queries stay in the RdfRepository
class. This increases the cohesion of RdfRepository.

Some more work is needed to cleanup the tests (and increase them since the
code is now easier to test). This will be done is subsequent patches.

Change-Id: Ib5fc41c43d860dc3c0fdeb0eca41d673abc25f1a