Enforce naming rule for unit tests.
The configuration of randomizedtesting plugin only runs tests that end
in "UnitTest" or "IntegrationTest". This is somewhat confusing and we
had quite a few test classes named "FooTest", which were not run during
any test phase.

An ugly beanshell rule is added to validate the naming rule and thus
ensure that all tests are run.

Change-Id: If6b6fbb0cb3d8180cc83dac75aad418d5c38cdb8