Allow continuation across empty results
Due to miser mode [1], the MediaWiki API may return empty results before
the full results is actually exhausted, and we may need to follow
continuation across any number of empty results before some more actual
results are returned: for example, in one particularly extreme case, [2]
is the first in a series of 197 empty responses before some more results
are returned in [3]. This patch enables MWAPI to continue across empty
results; however, this continuation must be limited to some degree in
order to avoid wasting query execution time and API resources on
following continuations of truly empty results forever.

The previous limit of MWAPI, which either caps the number of results or
disables continuation, is replaced with three, independent limits: one
on the number of results, one on the number of overall continuations,
and one on the number of consecutive empty continuations. The
configuration of the limits is backwards compatible: the limit on the
number of results uses the same predicate, and specifying it as "once"
is now equivalent to setting the limit on continuations to one. All
three limits are capped, to 10000 results or 1000 continuations by
default, but this can be changed using Java properties (the name of the
property for the results limit is likewise backwards compatible).

By default, the limits on the number of results and the number of
continuations are the same as their caps (10000 and 1000) if the query
does not specify the limit, but the limit on empty continuations
defaults to a much lower value: just 25 continuations. Query authors who
expect their generators to yield many empty pages should explicitly
specify a higher value, but for most users this lower limit will be more
than sufficient while avoiding excessive API load.

The following query demonstrates the syntax:

    SELECT * WHERE {
      SERVICE wikibase:mwapi {
        bd:serviceParam wikibase:endpoint "www.wikidata.org";
                        wikibase:api "Generator";
                        wikibase:limit "1000";
                        wikibase:limitContinuations "100";
                        wikibase:limitEmptyContinuations "50";
                        mwapi:generator "allrevisions";
                        mwapi:garvuser "Nikki";
                        mwapi:garvnamespace "146";
                        mwapi:garvlimit "max".
        ?title wikibase:apiOutput mwapi:title.
      }
    }

(This is the same API query which will, with sufficiently high limits on
continuations and empty continuations, yield the 197-empty-responses
series mentioned above.)

[1]: https://www.mediawiki.org/wiki/Manual:$wgMiserMode
[2]: https://www.wikidata.org/w/api.php?action=query&generator=allrevisions&garvnamespace=146&garvuser=Nikki&garvlimit=max&continue=garvcontinue%7C%7C&garvcontinue=20180731194922%7C718969437
[3]: https://www.wikidata.org/w/api.php?action=query&generator=allrevisions&garvnamespace=146&garvuser=Nikki&garvlimit=max&continue=garvcontinue%7C%7C&garvcontinue=20180531133552%7C687936602

Bug: T209034
Change-Id: I4ec2617565be11144d5d9d6fb20b7ace913bfa89