Scripts cleanup - fail on error, no-progress curl
There is no point for these scripts to silently ignore an error
and continue executing.  Also, adding `set -e` allows the caller
of the script to know that something crashed.

Also, make curl not output progress -- it mostly just
adds tons of junk in the logs.

Lastly, fixed a mixture of tabs and spaces in some files.

Change-Id: Id8a493fb5b6953b751c843879562f086efd06f50