Implement more cache-friendly Wikibase fetch strategy
If the change is old (currently > 3 hours) then we try to
fetch the most recent result overriding the cache, as it's likely
that there's more than one edit in this time period.
If the change is recent, then we fetch specific revision that
came in the change. Better mode would be "this revision or later" but
Wikidata does not implement such scheme yet.

Bug: T217897
Change-Id: I33b6acf57f20a6f384804f75077acbeed7f68597