Refactor the various URIs scheme to be more symetric.
We now have a reuseable SingleUrisScheme, which has 2 main instances,
one for wikidata and one for commons. They are composed with
CompoundUrisScheme, which exposes the fallback mechanism, independantly
from either wikidata or commons. It is (in my not so humble opinion)
now easier to understand how the fallback works.

As usual: prefer composition over inheritance.

Bug: T229608
Change-Id: Icccea41f5a2e1e80b92ba70b4b37af1c6d4ff1ab