Start dropping requests when load is too high.
This is a fairly aggressive approach, but we are living in interesting
times. This adds a filter that monitor system load and start dropping
requests with an HTTP 503 if load is above a certain threshold (80% of
the number of processors by default).

It is configured to only throttle requests that have a X-BIGDATA-READ-ONLY
HTTP header, so it let's the updater through.

Bug: T238229
Change-Id: Ib5c43d4d805008f9743ce2ef5498d90191e1aecc