Optimize the munge process
make it async by running:
- the RDFParser
- the munger
- the RDFWriter
in separate threads exchanging data using a blocking queue.

Cleanup the Munge class:
- remove the OutputPicker abstraction level.
- extract EntityMungingRdfHandler and add some unit tests

The speed gain to expect is around 2.5x the previous synchronous
approach. The RIOParser being the slowest component it might be
hard to optimize further since the entity delimitation relies on
it.

Change-Id: I30894dda4abf925e44309d0442b53d1c0894f170