Move entity statements without rank detection...
...out of the Munger

Added in I15b4c49cf9acf91e8d19f5e65260fa986bf8a484 as part of
an investigation done in T203646. This ticket is closed
so I assume that detecting these is no longer required.
We now run this detection only when lauching the Updater
with the verify option.

Change-Id: I015e1c3b6aa4123651a9dcb656354c6c8785e39d