Add data classes to represent the RDF diff stream
used by the streaming updater.
The parent class only declares a basic set of metadata:
- event metadata as defined in /fragment/common/1.1.0
- the entity ID
- the revision
- the original event time
- the sequence
- the number of sequences
- the operation type

For now only a DiffEventData class has been impletemented to support:
- import operation: provided the triples to link and the newly
  referenced shared triples
- diff operation: similar to import + triples to delete, and the
  unreferenced shared triples

Since the size of the largest entity in wikibase may surpass the
max size allowed by kafka a way to chunk an event is added.
The field sequence can be used to reconstruct a full event.

The various triples to manipulate are serialized as TURTLE for
now but could be changed to a more efficient serialization format
in the future using the mimeType added to the event.

Bug: T248464
Change-Id: Ic5848bee7794b6e0a794a8a15e8cfc333cfad4b3