Optimize wikibase:isSomeValue by inlining its namespace
Without inlining we might require to do a lookup in the forward index
on every value checked. This is probably too slow and might
cause timeouts on queries that previously worked with isBlank().
Use inlining capabilities of blazegraph to inline such values,
the prefix being part of the vocabulary should be fast to check.

Change-Id: I0488e1d544b1d278237743cdd61d1d0b9dc90eac