Set prefix file through runBlazegraph.sh
SEfC requires different namespace prefixes to be used
by Blazegraph, so we need a way to customize those. Prefix
configuration has been added and a parameter to set the file.
This allows for prefixes in query service to be consistent
with Commons dumps. Question remains if user prefer that,
or prefixes consistent with WDQS.

Bug: T251490
Change-Id: I3cce0df28cde1120b9ecafdf036bc6dee7c35b85