Switch GenerateEntityDiffPatchOperation to RichAsyncFunction
the async operators when used with a parallelism of 2 and a thread
pool of 30 (60 pending tasks) is able to backfill using default buffer
sizes emitting between 300 to 400 msg/sec while keeping checkpoint
times under 3 minutes (2 mins avg).

The timeout for the async wait operator has been set fairly high to
let the WikibaseRepo fails and retry on its own (hitting a timeout on
the operator fails the pipeline).

Added additional fixes:
- switch to kafka-client 2.4.1
- use flink's JavaSerializer

Bug: T258684
Change-Id: Ia17c8066dba807ccd65a4fefb2c2626bd0969412