Allow event sender if header overwrite
While normally options defined in web.xml can be overwritten,
there is no way to nullify any of them, so it was not possible to
remove the if header funtionality for event sender. This patch
checks if header name value is empty and works the same way as if
it was null.

Bug: T259637
Change-Id: I9500c1a6f56ff5efe8f664e27aaeb12eeb8a92be