Module reorganization for streaming updater
To make working with streaming updater code
easier this extracts the code from "tools"
module that is directly responsible for
consuming events and the code that is common
for both consumer and producer.

All of streaming updater modules depend on tools.
Extracting the code that they depend on from tools
would be beneficial, but this shouldn't be done in a
single step.

Bug: T251096
Change-Id: I77b9173bf53991889946828da66205c3cd8f82fd