Better resource handling for mwapi calls
pass the timeout to the request previously it was only used to quit
the awaitCondition on listener.get() without actually telling the http
client what timeout to use for the request itself.

Add more logging as the current logging was not very useful, when the
bug occurs the log line "MWAPI RESPONSE" is not visible suggesting
that an error occurred but no error is visible in the logs nor the
query output.

Properly close the listener.getInputStream() as suggested in the doc
which was left open.

Bug: T263952
Change-Id: I8b4b7b8385056bf852f7a82f23359e338e30fdfa