Reclassify sitelinks after diffing
Eventhough sitelinks are not supposed to be shared by muliple entities
their model allows it and only unicity check done in wikibase ensures
that they cannot be referenced from mutiple entities.
In the pipeline it might happen that links are renamed and then moved
to another entity during a merge operation. The proper ordering of this
operation is unfortunately not detectable as it happens at the same
time.
The goal of this patch is to relax the checks done while accumulating
patches so that if two entities try to add the same sitelink they
are allowed to do so (moving the sitelink blocks to the linkedshared
section).
But also in some case to the unlinkedshared section because we do not
know if this sitelink will be used.
The only moment we are confident that we can remove a sitelink block
is when it stays within the same entity (rename) in which case we
leave it in the removed section.

Bug: T266751
Change-Id: I537589f4a2b5db37a1e77b4090fc6ab1c81e60db