Unify re-ordering and decide mutation operation
The idea is to determine whether an event might require
buffering because another event that have not been received
might be required.

Use the parent revision to determine whether this is the
case or not. Delete/Undelete events are always buffered.

The expectations is that most events are well ordered (>95%).

If this is proven to work well we might want to consolidate
all of this and remove the EventReorderingWindowFunction
and properly merge both ReorderAndDecideMutationOperation
with DecideMutationOperationMapFunction so that state access
are more optimized.

Bug: T267029
Change-Id: I1f2c3ef1627251037909752a59ebb3849854b949