rdf-spark-tools: Use / as separator in partition args
When hive reports partitioning, such as through the `show partitions`
query, the partitioning is reported as k1=v1/k2=v2/.... Similarly the
python version of this, in HivePartition in wikimedia/discovery/analytics,
separates the kv pairs with /. Be consistent and use the same
separator here.

Checked with dcausse, the usage of this is currently ad-hoc so there
shouldn't be any specific pieces that must be uploaded.

Change-Id: I4f1d6d26b2089a742874ce00706012dd5e819357