Add custom serializers for stored objects
- InputEvent: serialized in the state with ReorderAndDecideMutationOperator
- MutationOperation: serialized by the async operator

Reason to use custom serializers is to support schema upgrades with
scala case classes.
Case classes are handy in the pipeline as it relies on scala pattern
matching. Schema upgrades allows to upgrade the pipeline without
relying on the state-processor API.
Other strategy like avro or other serializations engine supporting
schema evolution would require additional transformation to build the
scala classes so that it seems that a custom serializer is probably
the most straightforward way of handling this.

While using the state-processor is an OK solution it would require us
to inspect the state of the async operator (which we do not own) and
write the pending stream records in a serialized form that supports
schema upgrade anyways.

The new code is still guarded by a feature flag to ease the upgrade
to this new serialiazers as they are not compatible with flink's case
class serialiazers. This should become the new defaults soon.

Bug: T283599
Change-Id: I71918dc6fb216daa63887acae57e0e846929db51