Add option to use page id as entity id
When ingesting events we made the assumption that the page title can
be munged into the entity id, which was true for wikidata. For commons
the page title is the related file, the entity id has to be derived
through the page id.

To support split out an EntityResolver concept in IncomingStreams and
vary the implementation between traditional and using the prefixed
page_id depending on configuration.

Change-Id: If9e901bc1f0364304caa4a5db081822a5b014391