UrisSchemeFactory: Clarify when we mean wikidata and when we mean wikibase
The name wikibase seems to be used interchangably with wikidata in a
variety of the code. In some cases this is simply not the case though.
Here the UrisScheme returned is highly specific to either wikidata,
commons, or both. Make that clear by not calling things wikibase.

Avoids changing the system parameters referenced, assuming those are
also in a config file somewhere and needs updates in multiple places.

Change-Id: I6a88c5dc6c6089c8eb1acbd0f550f9330db05a5f