mw-oauth-proxy: Replace access token with stateless JWT
To support multi-instance deployment the access tokens need to
validate on any instance. Support this by issuing signed JWT tokens
that can be validated with only a shared secret. This has the downside
of granting irrevocable access tokens. We expect that a short timespan
on the tokens (2 hour default) should be sufficient for our use case
as this currently fronts a read-only service.

Change-Id: Ie5f2ea44fa733a2ebf1bd307048f2db404ef2bfe