blazegraph: Optionally include user identity in query events
To facilitate blocking in wcqs we need to know which user the requests
are coming from. On wcqs this information is available in the request
cookies, as long as the shared secret is the same value. Add a new
Filter that provides HttpServletRequest::getRemoteUser when provided
by jwt, and adjust the QueryEvent for schema 1.2.0 which has a place
for the username.

Bug: T293462
Change-Id: I3871084cd73262a63c142af80c73dae33b81ab5b