producer: Wait between retry attempts when fetching entity
We are seeing failures due to 404's for entities that do have content,
but the producer is racing with sql replication and sometimes wins.
Replace the fast-retry handling that was here with a delayed retry
mechanism which will, overall, increase the amount of time we will
wait for external replication to catch up before declaring it actually
a failure.

Bug: T279698
Change-Id: I5d70674af44d794f0c428071f6c010f03206be4f