producer: add support for reconciliation
Add a new optional input stream able to read reconcile events.

Pick the most recent revision from found either in the state or the
reconcile event and ship a reconcile event if the revision is
considered as present or a delete otherwise.  When revisions are
identical but one says the revision exists and the other the opposite
trust the reconcile event.

Changes to serialization formats are required:
- InputEventSerializer: to add support for ReconciliationInputEvent
- MutationOperationSerializer: to add for the Reconcile mutation
  operation

Json output for failed operations side-output slightly changes as well
to carry out the nature of the "reconcile" operation (see dependent
patch).

Bug: T279541
Depends-On: I571e22a58773e22232d2f0fd26e978cfff8c40b8
Change-Id: I3963f94b88733b8e6a11c4df5a18a75a0db4314e