Allow configuring the acceptable lag MW can have
The hard-coded value of 10 sec sounded correct for wikidata but data
shows that we are getting a non-negligible number of failures on
commons (20ish/day). This probably explains (at least partially)
why we get more detected inconsistencies from the consumer side on
wcqs.

Add a new config option so that this value can be increased a bit on
commons.

Change-Id: I37cc36b5a15b916a55f08e005bb56165791e8cb2