Productionize subgraph analysis metrics
Subgraph query mappings and metrics need to be executed and
evaluated over a longer period of time. This patch organizes all
such mappings and metrics, and saves them as tables for further use.
It extracts:
- List of subgraphs, Subgraph-Item mapping, Subgraph-Triple  mapping
- Subgraph-Query mapping
- Subgraph metrics (size, predicates, etc)
- Subgraph query metrics (query count, time, services etc)

See docs/spark_transform.md and docs/spark_metrics.md for dataflow
and output details. See docs/schema for schema of all the output tables.

Bug: T303831
Change-Id: Ia6a713db592894990bfe8c645049e08b00c61a1b