Reorganize rdf-spark-tools submodule
rdf-spark-tools submodule is refactored now to contain the following
package structure:

spark
- transform
  - structureddata
     - dumps
  - queries
    - sparql

This was done based on what is being performed, here: transformation
on data. The data is either structureddata (wikidata or WM commons)
or queries related to these. Inside of structureddata.dumps dumps are
being generated, and inside of queries.sparql wdqs SPARQL queries are
being processed.

This is done before adding subgraph-analysis-productionization code to
better settle those and other future codes in proper places.

Bug: T303831
Change-Id: I77f42cb86cd5018a0154ffd541eac0c93098eead