Unit tests for subgraph query mapping
SPARQL queries are mapped to the identified wikidata subgraphs
based on several criteria (items, properties usage etc) to
estimate which subgraphs are accessed by these queries.
This patch contains tests that use a small set of queries
on a small wikidata subset to identify which subgraphs are
accessed by these queries.

Bug: T303831
Change-Id: Ic6eeccec8ec542d8b04d4a90875baca5329755eb