Separate logging deps into dedicated package
We have a variety of logging dependencies that are duplicated between
mw-oauth-proxy and blazegraph. The end result today is that we still
have incomplete logging output - we get application specific logging
from mw-oauth-proxy and blazegraph, but requests that jetty kills
don't get any mention in logging.

The problem is that so far we have only ever configured individual
logging within the wars, leaving jetty with its StdErrLog which gives
minimal output on stderr. To resolve pull the logging dependencies out
into their own module and provide that to jetty on startup.

Bug: T306899
Change-Id: I7df40a4a0689c61c14d5774ebadf42cbb02b7c66