Fix runBlazegraph.sh detection of JETTY_RUNNER
Unclear why, in manual testing with various bash version this always
works, but the `(jetty-runner*.jar)`  syntax doesn't seem to
consistenly work when starting the systemd unit. Change to `$(echo
jetty-runner*.jar)` which makes things more explicit and less magical.
This will now expand to all matches, instead of only the first match,
but in practice there shouldn't be a second jetty-runner jar to match
against.

Change-Id: I09bf44b622b92317362f5d673d59fa2f59cefdac