mw-oauth-proxy: Rename accessTokenDurationHours to accessTokenDurationMinutes
A duration of hours might be convenient for reading, but it makes
testing difficult as the application can't be configured to forget
quickly while manually testing. This has never been set to a value
other than the default in production, so change the configuration
value from hourly resolution to minutes.

Additionally set the duration to 2 minutes in the test environment to
simplify testing.

Change-Id: I191bff868331e0e3674deb7ed58a1a5dca116469