mw-oauth-proxy: Update nginx test integration to work with POST
Adjusts the nginx configuration until it worked with POST requests.
What was happening with POST requests is that nginx would pass along
the Content-Length header to the internal requests but it didn't pass
along the body.  Fix by explicitly clearing the Content-Length header.

It's not clear why `proxy_pass_request_headers off` doesn't work on
it's own. Per the documentation that should prevent nginx from passing
along the headers, but it seems to stop all the headers except
Content-Length which must have separate handling.

The same fix needs to be applied to production instances.

Bug: T306899
Change-Id: I4657ba699caf78d5834b88615792f80bd422443a