mw-oauth-proxy: Set oauth cookie expiration in far future
This was using NewCookie.DEFAULT_MAX_AGE, which results in a cookie
that expires when the browser session closes. What we actually want
is for these cookies to last indefinately. The cookie spec doesn't
allow for that, we have to specify an amount of time, so set it to
approximately the number of days in ten years.

Change-Id: I122fd7455e4e10f458a48fe1d6789d80f6e3a14a