streaming-updater: specifically detect missed undelete events
The newer_revision_seen inconsistancy did encompass quite a few scenario
and notably was mixing some expected conditions (backfilling) and some
unexpected ones (missed undelete events).
Be explicit about this and add a new inconsistancy type.
Workaround this on the UdpaterReconcile spark job to only cover
newer_revision_seen emitted on top of a state that identifies a deleted
entity. (Note that this workaround can be removed once the
streaming-updater stops producing those).

Bug: T329089
Change-Id: Ib2952cd1e36df82110bc2849f6820cd2e452e074