Monitor read queries via JMX and QueryEventSenderFilter
blazegraph can report the number of queries it has started/done/failed
but this includes also update queries.
It might be interesting to monitor exlusively "read-only" queries which
are generally the ones sent by users. This could help to determine if a
particular node is actually serving user traffic or not.

Bug: T331405
Change-Id: I1003efeb2f8ac92af6f8fa6a1cbeb04b8f277033