UpdaterReconcile should be able to deal multiple DC in one run
Running the job per DC forces us to create the airflow dynamically
in a loop.
This is error prone and highly discouraged.
We should prefer a single run approach. Allow the job to run over data
that span multiple DC with the ability to construct the source_tag
based on the data it reads.
The pattern is $DC$ which will get replaced at runtime by the DC seen
in the rows.

Change-Id: I31bf81ebdc99903a13a9e6e534ba576ea3f2b9bb