Model EntityId as a specific class
We had an error where the entity id of `Main Page` made it through the
streaming updater, into the reconcile topics, and caused the
reconciliation phase to get all the way to querying wikidata about it.
The mediawiki api reported failed with the error `No such pageid: ain Page`.

To get errors closer to the source of the problem, instead of deep
within the system, create an EntityId class that wraps the string when
it comes into the system and will fail when the id isn't a proper
entity id. This hopefully also makes various bits of code a little
easier to follow as we now have an explicit EntityId type instead of
arbitrary String instances.

This avoids changing the case classes the streaming updater uses
to pass data around. I'm not sure how all the serialization works
there and expect we need to be careful to ensure it would transition
seamlessly between String and EntityId values when deployed.

Change-Id: Id57e5e0b5f99635233e3e5fefb4fe6bdfdab8da0