Allow reconciling a deletion without knowing its revision
The streaming-updater-producer flink job is heavily revision based to do
its consistency checks. Sadly in some conditions it is hard to determine
the revision to work on and might require special MW permissions to
extract this information. This is the case for deleted items where the
sole information we can publicly infer is that the item was deleted.
For such case we allow sending a reconcile event with a revision id set
as 0 to bypass any revision consistency check.

Update the dist build to include the reconcile_items.py script that can
be used to manually reconcile a list of items.

Bug: T342593
Change-Id: I8bc7b46da52a1734fd5a9fd6bb34fc13cafeb6aa