Add an option to use newer KafkaSource/Sink flink APIs
The use_new_flink_kafka_api option will tell the pipeline to use
KafkaSink/Source instead of FlinkKafkaConsumer and Producer.
Operator names have been kept if still using the old APIs but changed
if using the new API. This allows to load a savepoint with the new API
re-using offsets stored in kafka but discarding old operator states
with allowNonRestoredState.

Bug: T326914
Change-Id: I762d696cd7cef5f8b8ac26a2bd8da21989bcc9a4