Kafka sources should rely on existing group offsets
when not found in the savepoint state. The job will fail if no group
offsets are available.
As opposed to the deprecated FlinkKafkaConsumer which defaults to
GROUP_OFFSETS KafkaSource defaults to earliest which is never something
we want with this job, we always start from an existing state and its
corresponding offsets.

Bug: T326914
Change-Id: Ia5da25d4880151d88b9149efdd2cd0577d024966