spark queries processor: Force a shuffle on input
We are having a problem where this job runs within a single executor
and requires, at times, 12g of memory overhead to run to completion.
Attempt to get memory usage under control by forcing a shuffle when
reading the data, which will allow spark to distribute the tasks
to multiple executors. We already have an option to set the number
of output partitions, without that we would end up writing 200
partitions.

Bug: T347333
Change-Id: I36fe2c929f2de1112dc034c0a4e1a3d89c5f0618