spark query extractor: Generalize partitioning
In a prior patch, a9551af415b, partitioning was updated to verify it
would fix an issue with the jobs running out of memory. This has
worked, but left the configuration a bit confusing. Clean things up
by making explicit options for input and output partitioning, and
leaving a BC option for the old way.

This is only necessary here, rather than in general spark jobs, because
this job doesn't involve any shuffles but does a significant amount of
work per row. Spark generally is a little more specialized towards
simple per row operations like you see in SQL, but works acceptably
for this use case as well.

Change-Id: I2f926f356cdfd6918b9b3bb338f406835a739e5b