subgraph mapper: Write and reuse intermediate results
This job has results that are both saved to disk and used later in the
computation. Avoid double computation by reading the data we wrote
to disk back out and using that for future computations.

This additionally has the side benefit of clearly separating the
execution into three distinct stages, ordered by the order we write
them to disk. This additionally makes it more obvious what each stage
of execution is doing in the execution graph.  Essentially by
computing from the values stored on disk the execution graph is
smaller and has better labeled inputs. Where it once had a multi-step
transformation it now simply says the table name.

Change-Id: I3c667de6e0b56d24686ae2e22ce3634182a3eeb7