Use LATEST offset reset strategy
Initial run is apparently not the sole condition in which the offsets
might have to be reset.
A problem occured few days after the switch to the new KafkaSource
implementation. The application failed to restart with
NoOffsetForPartitionException: Undefined offset with no reset policy for
partitions. So it appears that even if a consumer group is bound to a
topic kafka might forget its offset if there are no activity on it:
https://issues.apache.org/jira/browse/KAFKA-4682.
Use LATEST because it was the default used by the previous
FlinkKafkaConsumer (that relies on default kafka auto.offset.reset).
This is not ideal as it's unclear if some messages might get lost on
such idle stream when events are starting to flow again.
Using EARLIEST sounds dangerous as well esp if the offset retention is
shorter than the topic retention, the stream after going idle might be
reset to EARLIEST after the offsets are gone causing many events to be
reprocessed.
I'm not clear if there are ways to solve this problem without either:
- have almost infinite retention of offsets (offset.retention.minutes)
- upgrade to kafka >= 2.1
- do not have idle streams for long periods (longer than
  offset.retention.minutes). Canary events everywhere?

Bug: T349147
Change-Id: I89f81a8eb0ff2f8c97cd81a46a6f0c718c347f63